+++ src/components/NormalizedStackedBarChart/components/BarLabel/tests/BarLabel.test.tsx
@@ -21,6 +21,7 @@ describe('<BarLabel />', () => {
             },
           }}
           orientation="horizontal"
+          labelPosition="top-left"
         />,
       );