+++ src/components/NormalizedStackedBarChart/stories/NormalizedStackedBarChart.stories.tsx
@@ -26,6 +26,9 @@ export default {
     },
     orientation: {description: 'Determines the orientation of the chart.'},
     size: {description: 'Determines the width of the chart.'},
+    labelPosition: {
+      description: 'Determines the position of the labels.',
+    },
     theme: THEME_CONTROL_ARGS,
   },
 } as Meta;
@@ -61,6 +64,7 @@ const defaultProps = {
   ],
   orientation: 'horizontal' as 'horizontal',
   size: 'small' as 'small',
+  labelPosition: 'top-left' as 'top-left',
 };
 
 export const Default = Template.bind({});